// src/models/ProductsModel.ts
export interface IPropProducts {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: [],
    brand: string;
    sku: string;
    weight: number;
    dimensions: {
        width: number;
        height: number;
        depth: number;
    }
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: [
        {
            "rating": number;
            "comment": string;
            "date": string;
            "reviewerName": string;
            "reviewerEmail": string;
        },
        {
            "rating": number;
            "comment": string;
            "date": string;
            "reviewerName": string;
            "reviewerEmail": string;
        },
        {
            "rating": number;
            "comment": string;
            "date": string;
            "reviewerName": string;
            "reviewerEmail": string;
        }
    ],
    "returnPolicy": string;
    "minimumOrderQuantity": number;
    "meta": {
        "createdAt": string;
        "updatedAt": string;
        "barcode": string;
        "qrCode": string;
    },
    "images": [];
    thumbnail: string;
}